worker/requirements.txt
fastapi
uvicorn
google-cloud-storage
google-cloud-firestore
# force rebuild v2